+++ src/boards/arcade_feather_nrf52840_express/board.h
@@ -22,8 +22,8 @@
  * THE SOFTWARE.
  */
 
-#ifndef _FEATHER_NRF52840_H
-#define _FEATHER_NRF52840_H
+#ifndef _ARCARDE_FEATHER_NRF52840_H
+#define _ARCARDE_FEATHER_NRF52840_H
 
 #define _PINNUM(port, pin)    ((port)*32 + (pin))
 
@@ -62,8 +62,9 @@
 
 //------------- UF2 -------------//
 #define UF2_PRODUCT_NAME      "Adafruit Feather nRF52840 Express"
-#define UF2_VOLUME_LABEL      "FTHR840BOOT"
+#define UF2_VOLUME_LABEL      "ARCADE-N4"
 #define UF2_BOARD_ID          "nRF52840-Feather-revD"
 #define UF2_INDEX_URL         "https://www.adafruit.com/product/4062"
 
-#endif // _FEATHER_NRF52840_H
+#endif // _ARCARDE_FEATHER_NRF52840_H
+