+++ src/boards/clue_nrf52840/board.h
@@ -46,18 +46,11 @@
 #define BUTTON_2              _PINNUM(1, 10)     // right button
 #define BUTTON_PULL           NRF_GPIO_PIN_PULLUP
 
-/*------------------------------------------------------------------*/
-/* UART (only used by nRF52832)
- *------------------------------------------------------------------*/
-#define RX_PIN_NUMBER         _PINNUM(0, 09)
-#define TX_PIN_NUMBER         _PINNUM(0, 10)
-#define CTS_PIN_NUMBER        0
-#define RTS_PIN_NUMBER        0
-#define HWFC                  false
-
-// Used as model string in OTA mode
+//--------------------------------------------------------------------+
+// BLE OTA
+//--------------------------------------------------------------------+
 #define BLEDIS_MANUFACTURER   "Adafruit Industries"
-#define BLEDIS_MODEL          "Clue nRF52840"
+#define BLEDIS_MODEL          "CLUE nRF52840"
 
 //--------------------------------------------------------------------+
 // USB
@@ -67,9 +60,9 @@
 #define USB_DESC_CDC_ONLY_PID  0x0071
 
 //------------- UF2 -------------//
-#define UF2_PRODUCT_NAME   "Adafruit Clue nRF52840"
+#define UF2_PRODUCT_NAME   "Adafruit CLUE nRF52840"
 #define UF2_VOLUME_LABEL   "CLUEBOOT"
-#define UF2_BOARD_ID       "nRF52840-Clue-revA"
-#define UF2_INDEX_URL      "https://www.adafruit.com/"
+#define UF2_BOARD_ID       "nRF52840-CLUE-revA"
+#define UF2_INDEX_URL      "https://www.adafruit.com/product/4500"
 
 #endif // _CLUE_NRF52840_H