+++ changelog.md
@@ -1,5 +1,12 @@
 # Adafruit nRF52 Bootloader Changelog
 
+## 0.3.2 - 2020.03.12
+
+- Make sure all pins are in reset state when jumping to app mode.
+
+## 0.3.1 - 2020.03.05
+- Correct LED1 for Feather Sense and change volume name to FTHRSNSBOOT.
+
 ## 0.3.0 - 2020.01.13
 
 - Upgrade nrfx to v2 for supporting future nrf mcu such as nrf52833, nrf5340
@@ -29,7 +36,7 @@
 - NFC pins are forced to GPIO mode by bootloader
 - Added Metro nRF52840 Express VID/PID
 - Enhance board management
-- Added electronut/papyr_support 
+- Added electronut/papyr_support
 
 ## 0.2.9
 
@@ -45,7 +52,7 @@
 - Fixed PWM psel[1] is not reset
 - Fixed #41 move RXD, TXD into board header
 - Added Metro nRF52840 Rev A
-- Fixed #40 OTA issue with BLE_GAP_EVT_PHY_UPDATE_REQUEST e.g connecting with iPhone X  
+- Fixed #40 OTA issue with BLE_GAP_EVT_PHY_UPDATE_REQUEST e.g connecting with iPhone X
 
 ## 0.2.6
 
@@ -63,4 +70,3 @@
 - Fully support Feather nRF52840
 - Update bootloader with new led pattern
 - Fix #203: return software timer handle
-